TypeScript Assignment
grocery.ts is a typescript file that instantiates a grocery class that has the following attributes: name, quantity, and perishable. grocery.ts creates an array full of grocery items and displays the array into the groceries.html file. groceries.html calls upon the typescript, which is compiled into a JS file. Should correctly display the array full of groceries on a web page.